Move test on chunk_list_ in test_api_data as fonction move to #194
